+++ src/Instrumentum_sanae_doctrinae/web_scraping/monergism/monergism_scrap_general_information.py
@@ -580,7 +580,7 @@ class MonergismScrapGeneralInformation_ALL(http_connexion.ParallelHttpConnexionW
                                        my_constants.MONERGISM_NAME,
                                        my_constants.ELABORATED_DATA_FOLDER,
                                        browse_by_type,
-                                       my_constants.SPEAKER_TOPIC_OR_SCRIPTURE_LISTING_FOLDER,
+                                       my_constants.GENERAL_INFORMATION_NAME,
                                        my_constants.get_default_json_filename(0))
         
         input_root_folder = os.path.join(root_folder,
