+++ tests/sermonindex/test_sermonindex_audio_sermon_scrap_work.py
@@ -1,0 +1,25 @@
+import asyncio
+import sys
+from Instrumentum_sanae_doctrinae.web_scraping import my_constants
+from Instrumentum_sanae_doctrinae.web_scraping.sermonindex import sermonindex_audio_sermon_scrap_work 
+
+
+def test_audio_sermon_get_scripture_work():
+    root_folder ='D:/projet_github/FOR GOD/Scraping general/test_folder'
+    material_folder = my_constants.SERMONINDEX_AUDIO_SERMONS_ROOT_FOLDER
+    browse_by_type = my_constants.SPEAKER_NAME
+    ob = sermonindex_audio_sermon_scrap_work.SermonIndexScrapScripturesWork_ALL(
+        root_folder,
+        material_folder,
+        browse_by_type,
+        overwrite_log=True
+    )
+    #print(ob.__dict__)
+    asyncio.run(ob.download(2))
+    #ob.update_downloaded_and_to_download()
+    #ob.write_log_file()
+    
+if __name__ == "__main__":
+    if sys.platform == "win32":
+        asyncio.set_event_loop_policy(asyncio.WindowsSelectorEventLoopPolicy())
+    test_audio_sermon_get_scripture_work()
